Rename SplitSdk to SplitFactory function
