+++ src/__tests__/useSplitManager.test.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { act, render } from '@testing-library/react';
 
 /** Mocks */
-import { Event, mockSdk } from './testUtils/mockSplitSdk';
+import { Event, mockSdk } from './testUtils/mockSplitFactory';
 jest.mock('@splitsoftware/splitio/client', () => {
   return { SplitFactory: mockSdk() };
 });
-import { SplitFactory as SplitSdk } from '@splitsoftware/splitio/client';
+import { SplitFactory } from '@splitsoftware/splitio/client';
 import { sdkBrowser } from './testUtils/sdkConfigs';
 import { getStatus } from '../utils';
 
@@ -17,7 +17,7 @@ import { useSplitManager } from '../useSplitManager';
 describe('useSplitManager', () => {
 
   test('returns the factory manager from the Split context, and updates when the context changes.', () => {
-    const outerFactory = SplitSdk(sdkBrowser);
+    const outerFactory = SplitFactory(sdkBrowser);
     let hookResult;
     render(
       <SplitFactoryProvider factory={outerFactory} >
